app/database.py
import motor.motor_asyncio
from dotenv import load_dotenv
import os

load_dotenv()

MONGO_DETAILS = os.getenv("MONGO_URL")

client = motor.motor_asyncio.AsyncIOMotorClient(MONGO_DETAILS)
database = client.habit_tracker
habit_collection = database.get_collection("habits")