requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
fastapi==0.136.0
h11==0.16.0
idna==3.11
motor==3.7.1
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
pymongo==4.16.0
python-dotenv==1.2.2
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0